@@ -1,21 +1,52 @@
 [project]
 name = "pysatl-criterion"
-version = "0.0.1"
+version = "0.0.1-alpha"
+description = "Package with statistic criteria for PySATL"
 authors = [
-    { name="Example Author", email="author@example.com" },
+    {name = "PySATL Team", email = "pysatl@yahoo.com"},
 ]
-description = "PYSATL Criterion"
+license = {text = "MIT"}
 readme = "README.md"
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
+requires-python = ">=3.9,<3.13"
+dependencies = [
+    "numpy>=1.25.1",
+    "scipy>=1.11.2",
+    "pandas>=2.2.1",
+    "typing-extensions>=4.12.2"
 ]
 
 [project.urls]
-Homepage = "https://github.com/alex98247/statistic-test"
-Issues = "https://github.com/alex98247/statistic-test/issues"
+Homepage = "https://github.com/PySATL/pysatl-criterion"
+Issues = "https://github.com/PySATL/pysatl-criterion/issues"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry]
+name = "pysatl"
+packages = [
+    { include = "pysatl" },
+]
+package-mode = true
+
+[tool.poetry.group.dev.dependencies]
+markdown = "3.7"
+mkdocs = "1.6.1"
+mkdocs-material = "9.5.44"
+mdx-truly-sane-lists = "1.3"
+pymdown-extensions = "10.12"
+jinja2 = "3.1.4"
+mike = "2.1.3"
+isort = "5.13.2"
+coveralls = "4.0.1"
+pytest = "8.3.3"
+pytest-cov = "6.0.0"
+pytest-random-order = "1.1.1"
+ruff = "0.7.4"
+pytest-mock = "3.14.0"
+pre-commit = "4.0.1"
+mypy = "1.15.0"
 
 [tool.isort]
 line_length = 100
