@@ -1,5 +1,4 @@
 from abc import ABC, abstractmethod
-from typing import Optional, Tuple, Union
 
 from numpy import float64
 
@@ -16,7 +15,7 @@ class AbstractStatistic(ABC):
         raise NotImplementedError("Method is not implemented")
 
     @abstractmethod
-    def execute_statistic(self, rvs, **kwargs) -> Union[float, float64]:
+    def execute_statistic(self, rvs, **kwargs) -> float | float64:
         """
         Execute test statistic and return calculated statistic value.
         :param rvs: rvs data to calculated statistic value
@@ -24,7 +23,7 @@ class AbstractStatistic(ABC):
         """
         raise NotImplementedError("Method is not implemented")
 
-    def calculate_critical_value(self, rvs_size, sl) -> Union[Optional[float], Optional[float64]]:
+    def calculate_critical_value(self, rvs_size, sl) -> float | float64 | None:
         """
         Calculate critical value for test statistics
         :param rvs_size: rvs size
@@ -32,9 +31,7 @@ class AbstractStatistic(ABC):
         """
         return None
 
-    def calculate_two_tailed_critical_values(
-        self, rvs_size: int, sl
-    ) -> Optional[Tuple[float, float]]:
+    def calculate_two_tailed_critical_values(self, rvs_size: int, sl) -> tuple[float, float] | None:
         """
         Calculate two-tailed critical values for test statistics
         :param rvs_size: rvs size
